+++ src/Propulsion.Kafka/Bindings.fs
@@ -1,7 +1,7 @@
 ﻿namespace Propulsion.Kafka
 
 open Confluent.Kafka
-open Jet.ConfluentKafka.FSharp
+open FsKafka
 open Serilog
 open System
 open System.Collections.Generic